Add user parameter to init.pp
This allow auth.json and config.json to be created under the same user
as in project/exec calls. Because auth.json and config.json is created
with permission of 700, by default, they will be created for root user
(the default user). When user parameter is specified in project/exec,
composer are executed under non-root user and will get access denied
error.

When using this parameter, composer_home has also be set to a directory
that the user can read from. e.g. /tmp or /home/USER